+++ app/src/main/java/labs/zero_one/roundo/MainActivity.kt
@@ -11,14 +11,10 @@ import android.provider.Settings
 import android.support.v4.content.ContextCompat
 import android.support.v4.graphics.drawable.DrawableCompat
 import android.support.v7.app.AppCompatActivity
-import android.support.v7.preference.PreferenceManager
 import android.view.View
-import android.view.ViewGroup
 import android.widget.Toast
 import kotlinx.android.synthetic.main.activity_main.*
 import kotlinx.android.synthetic.main.content_main.*
-import kotlinx.android.synthetic.main.dialog_dashboard.view.*
-import java.util.*
 
 /**
  * 主页面 Activity
@@ -29,7 +25,7 @@ import java.util.*
  * - [locationKitListener]
  * - [missionManager]
  * - [missionListener]
- * - [dashboardLayout]
+ * - [dashboard]
  *
  * ## 子类列表
  * - [AppRequest]
@@ -43,7 +39,6 @@ import java.util.*
  * - [onRequestPermissionsResult]
  *
  * ## 自定义方法列表
- * - [openDashboard]
  *
  * @author lucka-me
  * @since 0.1
@@ -53,7 +48,7 @@ import java.util.*
  * @property [locationKitListener] 位置工具消息监听器
  * @property [missionManager] 任务管理器
  * @property [missionListener] 任务消息监听器
- * @property [dashboardLayout] 仪表盘视图
+ * @property [dashboard] 仪表盘
  */
 class MainActivity : AppCompatActivity() {
 
@@ -136,7 +131,7 @@ class MainActivity : AppCompatActivity() {
                         this@MainActivity, R.string.mission_resumed, Toast.LENGTH_LONG
                     ).show()
                 } else {
-                    openDashboard()
+                    dashboard.show(this@MainActivity)
                 }
                 // Update location
                 if (locationKit.isLocationAvailable)
@@ -187,12 +182,12 @@ class MainActivity : AppCompatActivity() {
                     R.string.confirm,
                     negativeButtonTextId = R.string.dashboard_title,
                     negativeButtonListener = { _, _ ->
-                        if (!isDashboardShown()) openDashboard()
+                        dashboard.show(this@MainActivity)
                     },
                     icon = icon
                 )
 
-                updateDashboard(indexList.size)
+                dashboard.update(indexList.size)
                 progressBar.incrementProgressBy(indexList.size)
             }
 
@@ -203,14 +198,13 @@ class MainActivity : AppCompatActivity() {
                     R.string.mission_all_checked_message,
                     R.string.dashboard_title,
                     positiveButtonListener = { _, _ ->
-                        if (!isDashboardShown()) openDashboard()
+                        dashboard.show(this@MainActivity)
                     },
                     icon = getDrawable(R.drawable.ic_mission_all_checked)
                 )
             }
 
             override fun onTimeUpdated(pastTime: Long) {
-                updateDashboard(0)
                 progressBar.incrementSecondaryProgressBy(
                     (1.0 * pastTime
                         * missionManager.checkPointList.size / missionManager.data.targetTime
@@ -220,15 +214,13 @@ class MainActivity : AppCompatActivity() {
             }
 
             override fun onSecondUpdated() {
-                updateDashboard(0)
+                dashboard.update()
             }
 
     }
     private val missionManager = MissionManager(this, missionListener)
 
-    private lateinit var dashboardLayout: View
-
-    //private lateinit var backgroundMissionService: Intent
+    private lateinit var dashboard: Dashboard
 
     /**
      * 请求代码
@@ -262,7 +254,7 @@ class MainActivity : AppCompatActivity() {
         setContentView(R.layout.activity_main)
 
         // Initialize Dashboard
-        initDashboard()
+        dashboard = Dashboard(this, missionManager)
 
         // Handle the permissions
         locationKit = LocationKit(this, locationKitListener)
@@ -309,7 +301,7 @@ class MainActivity : AppCompatActivity() {
         buttonDashboard.setOnClickListener {
             when(missionManager.state) {
                 MissionManager.MissionState.Started, MissionManager.MissionState.Stopped -> {
-                    if (!isDashboardShown()) openDashboard()
+                    dashboard.show(this)
                 }
                 else -> {
 
@@ -427,193 +419,5 @@ class MainActivity : AppCompatActivity() {
             }
         }
 
-    }
-
-    /**
-     * 打开仪表盘对话框，若任务开始则显示仪表盘，否则显示未开始任务。
-     *
-     * ## Changelog
-     * ### 0.2.1
-     * - 将 [dashboardLayout] 独立，实现实时刷新
-     *
-     * @author lucka-me
-     * @since 0.1.14
-     */
-    private fun openDashboard() {
-        val isDash = PreferenceManager
-            .getDefaultSharedPreferences(this)
-            .getBoolean(getString(R.string.pref_dash_enable_key), false)
-        val dialogBuilder = AlertDialog.Builder(this)
-            .setTitle(if (isDash) R.string.dashboard_title_ee else R.string.dashboard_title)
-            .setIcon(if (isDash) R.drawable.ic_dash else R.drawable.ic_dashboard)
-            .setPositiveButton(R.string.confirm, null)
-        val dashboard = if (missionManager.state == MissionManager.MissionState.Started) {
-
-            initDashboard()
-            var checkedCount = 0
-            for (waypoint in missionManager.checkPointList) if (waypoint.checked) checkedCount++
-            updateDashboard(checkedCount, true)
-
-            dialogBuilder
-                .setView(dashboardLayout)
-                .setNegativeButton(R.string.dashboard_stop) { _, _ ->
-                    DialogKit.showDialog(
-                        this,
-                        R.string.alert_title, R.string.mission_stop_confirm_message,
-                        R.string.confirm,
-                        positiveButtonListener = { _, _ ->
-                            missionManager.stop()
-                        },
-                        negativeButtonTextId = R.string.cancel,
-                        cancelable = false
-                    )
-                }
-                .setOnDismissListener {
-                    quitDashboardParent()
-                }
-                .show()
-
-        } else {
-
-            dialogBuilder
-                .setMessage(R.string.dashboard_mission_stopped)
-                .setNegativeButton(R.string.dashboard_start) { dialog, _ ->
-                    dialog.dismiss()
-                    quitDashboardParent()
-                    val intent = Intent(this, SetupActivity::class.java)
-                    startActivityForResult(intent, AppRequest.ActivitySetup.code)
-                    overridePendingTransition(R.anim.slide_bottom_up, R.anim.slide_bottom_down)
-                }
-                .show()
-
-        }
-
-        // Set dialog style
-
-        dashboard
-            .getButton(AlertDialog.BUTTON_POSITIVE)
-            .setTextColor(ContextCompat.getColor(this, R.color.colorPrimary))
-        dashboard
-            .getButton(AlertDialog.BUTTON_NEGATIVE)
-            .setTextColor(ContextCompat.getColor(this, R.color.colorPrimary))
-    }
-
-    /**
-     * 初始化仪表盘
-     *
-     * @author lucka-me
-     * @since 0.2.1
-     */
-    private fun initDashboard() {
-        dashboardLayout = View.inflate(this, R.layout.dialog_dashboard, null)
-
-        dashboardLayout.missionProgressBar.progress = 0
-        dashboardLayout.missionProgressBar.max = missionManager.checkPointList.size
-        dashboardLayout.missionSequentialTitle.text =
-            getString(
-                if (missionManager.data.sequential) R.string.dashboard_mission_sequential_title_true
-                else R.string.dashboard_mission_sequential_title_false
-            )
-        dashboardLayout.missionSequentialText.text =
-            getString(
-                if (missionManager.data.sequential) R.string.dashboard_mission_sequential_text_true
-                else R.string.dashboard_mission_sequential_text_false
-            )
-        dashboardLayout.timeProgressBar.progress = 0
-        dashboardLayout.timeProgressBar.max = missionManager.data.targetTime
-        dashboardLayout.timeProgressBar.secondaryProgress = 0
-    }
-
-    /**
-     * 更新仪表盘
-     *
-     * @param [checkedCount] 新签到的任务点数量
-     * @param [force] 强制更新
-     *
-     * @author lucka-me
-     * @since 0.2.1
-     */
-    private fun updateDashboard(checkedCount: Int, force: Boolean = false) {
-
-        if (!isDashboardShown() && !force) return
-
-        val showSecond = PreferenceManager
-            .getDefaultSharedPreferences(this)
-            .getBoolean(getString(R.string.pref_display_show_second_key), false)
-
-        dashboardLayout.missionProgressBar
-            .incrementProgressBy(checkedCount)
-        dashboardLayout.missionProgressText.text = String.format(
-            getString(R.string.dashboard_mission_progress_text),
-            dashboardLayout.missionProgressBar.progress, missionManager.checkPointList.size
-        )
-
-        // Mission Time
-        val realPastTime = ((Date().time - missionManager.data.startTime.time) / 1000).toInt()
-        dashboardLayout.timeProgressText.text = String.format(
-            getString(R.string.dashboard_time_progress_text),
-            timeToString(missionManager.data.pastTime, showSecond),
-            timeToString(missionManager.data.targetTime, showSecond)
-        )
-        dashboardLayout.timeStartText.text = if (showSecond) {
-            String.format("%tT", missionManager.data.startTime)
-        } else {
-            String.format("%tR", missionManager.data.startTime)
-        }
-        dashboardLayout.timeRealPastText.text = timeToString(realPastTime, showSecond)
-        dashboardLayout.timeProgressBar.incrementProgressBy(missionManager.data.pastTime
-            - dashboardLayout.timeProgressBar.progress)
-        dashboardLayout.timeProgressBar.incrementSecondaryProgressBy(realPastTime
-            - dashboardLayout.timeProgressBar.secondaryProgress)
-
-        // Distance
-        dashboardLayout.distanceText.text = if (missionManager.data.distance > 1000) {
-            String.format(
-                getString(R.string.dashboard_distance_text_km),
-                missionManager.data.distance / 1000
-            )
-        } else {
-            String.format(
-                getString(R.string.dashboard_distance_text_m),
-                missionManager.data.distance
-            )
-        }
-
-    }
-
-    /**
-     * 将仪表盘从其亲视图移除，仅此才可以置入新的对话框视图中
-     *
-     * @see <a href="https://stackoverflow.com/a/28071422">Stack Overflow</a>
-     *
-     * @author lucka-me
-     * @since 0.2.1
-     */
-    private fun quitDashboardParent() {
-        if (isDashboardShown())
-            (dashboardLayout.parent as ViewGroup).removeView(dashboardLayout)
-    }
-
-    private fun isDashboardShown(): Boolean {
-        return dashboardLayout.parent != null
-    }
-
-    /**
-     * 将秒数转换成格式化字符串
-     *
-     * HH:mm:ss 或 HH:mm
-     *
-     * @author lucka-me
-     * @since 0.2.1
-     */
-    private fun timeToString(second: Int, showSecond: Boolean): String {
-        val hrs: Int = second / 3600
-        val min: Int = (second - hrs * 3600) / 60
-        return if (showSecond) {
-            val sec: Int = second - hrs * 3600 - min * 60
-            String.format(getString(R.string.format_time_sec), hrs, min, sec)
-        } else {
-            String.format(getString(R.string.format_time), hrs, min)
-        }
     }
 }